+++ cmd/app_workspace_action.go
@@ -6,7 +6,6 @@ import (
 	"github.com/urfave/cli/v2"
 	"gopkg.in/nullstone-io/go-api-client.v0"
 	"log"
-	"os"
 )
 
 type AppWorkspaceFn func(ctx context.Context, cfg api.Config, appDetails app.Details) error
@@ -17,8 +16,11 @@ func AppWorkspaceAction(c *cli.Context, fn AppWorkspaceFn) error {
 		return err
 	}
 
+	waitForLaunch := c.Bool(WaitForLaunchFlag.Name)
+	osWriters := CliOsWriters{Context: c}
+
 	return ParseAppEnv(c, true, func(stackName, appName, envName string) error {
-		logger := log.New(os.Stderr, "", 0)
+		logger := log.New(osWriters.Stderr(), "", 0)
 		logger.Printf("Performing application command (Org=%s, App=%s, Stack=%s, Env=%s)", cfg.OrgName, appName, stackName, envName)
 		logger.Println()
 
@@ -28,6 +30,10 @@ func AppWorkspaceAction(c *cli.Context, fn AppWorkspaceFn) error {
 		}
 
 		return CancellableAction(func(ctx context.Context) error {
+			if err := WaitForLaunch(ctx, osWriters, cfg, appDetails, waitForLaunch); err != nil {
+				return err
+			}
+
 			return fn(ctx, cfg, app.Details{
 				App:       appDetails.App,
 				Env:       appDetails.Env,