+++ cmd/find_app_details.go
@@ -5,7 +5,6 @@ import (
 	"github.com/nullstone-io/deployment-sdk/app"
 	"gopkg.in/nullstone-io/go-api-client.v0"
 	"gopkg.in/nullstone-io/go-api-client.v0/find"
-	"gopkg.in/nullstone-io/go-api-client.v0/types"
 )
 
 // FindAppDetails retrieves the app, env, and workspace
@@ -20,8 +19,12 @@ func FindAppDetails(cfg api.Config, appName, stackName, envName string) (app.Det
 	appDetails.App = application
 	appDetails.Env = env
 
-	if appDetails.Workspace, err = getAppWorkspace(cfg, appDetails.App, appDetails.Env); err != nil {
-		return appDetails, err
+	client := api.Client{Config: cfg}
+	appDetails.Workspace, err = client.Workspaces().Get(appDetails.App.StackId, appDetails.App.Id, env.Id)
+	if err != nil {
+		return appDetails, fmt.Errorf("error retrieving workspace: %w", err)
+	} else if appDetails.Workspace == nil {
+		return appDetails, fmt.Errorf("workspace %q does not exist", err)
 	}
 
 	if appDetails.Module, err = find.Module(cfg, appDetails.App.ModuleSource); err != nil {
@@ -31,19 +34,4 @@ func FindAppDetails(cfg api.Config, appName, stackName, envName string) (app.Det
 	}
 
 	return appDetails, nil
-}
-
-func getAppWorkspace(cfg api.Config, app *types.Application, env *types.Environment) (*types.Workspace, error) {
-	client := api.Client{Config: cfg}
-
-	workspace, err := client.Workspaces().Get(app.StackId, app.Id, env.Id)
-	if err != nil {
-		return nil, fmt.Errorf("error retrieving workspace: %w", err)
-	} else if workspace == nil {
-		return nil, fmt.Errorf("workspace %q does not exist", err)
-	}
-	if workspace.Status != types.WorkspaceStatusProvisioned {
-		return nil, fmt.Errorf("app %q has not been provisioned in %q environment yet", app.Name, env.Name)
-	}
-	return workspace, nil
 }